debian: always build GFDL docs
The original logic was to disable docs as standard but, on Ubuntu,
enable docs. Since we are concerning ourselves only with Ubuntu
packaging there is no need to keep the logic to enable or disable
documentation generation.

In doing this we take out some preprocessing of the postinst scripts
for gcc, g++, and gfortran packages.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>